requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
gunicorn==21.2.0 
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1